+++ src/ReinforcementLearning/Agents/ThompsonSamplingAgent.cs
@@ -179,7 +179,7 @@ public class ThompsonSamplingAgent<T> : ReinforcementLearningAgentBase<T>
     public override Vector<T> Predict(Vector<T> input) => SelectAction(input, false);
     public Task<Vector<T>> PredictAsync(Vector<T> input) => Task.FromResult(Predict(input));
     public Task TrainAsync() { Train(); return Task.CompletedTask; }
-    public override ModelMetadata<T> GetModelMetadata() => new ModelMetadata<T> { FeatureCount = this.FeatureCount, Complexity = (int)ParameterCount };
+    public override ModelMetadata<T> GetModelMetadata() => new ModelMetadata<T> { FeatureCount = this.FeatureCount, Complexity = ParameterCount };
     public override long ParameterCount => _options.NumArms * 2;
     public override int FeatureCount => 1;
     public override byte[] Serialize()
